+++ flashinfer/fused_moe/cute_dsl/b12x_moe.py
@@ -76,7 +76,6 @@ def b12x_fused_moe(
     activation_precision: str = "fp4",
     quant_mode: Optional[str] = None,
     source_format: str = "modelopt",
-    swiglu_limit: Optional[float] = None,
     apply_router_weight_on_input: bool = False,
 ) -> torch.Tensor:
     r"""Run fused MoE on SM120/SM121 using b12x CuTe-DSL kernels.
@@ -209,7 +208,6 @@ def b12x_fused_moe(
         activation_precision=activation_precision,
         quant_mode=quant_mode,
         source_format=source_format,
-        swiglu_limit=swiglu_limit,
         apply_router_weight_on_input=apply_router_weight_on_input,
     )
 
@@ -262,7 +260,6 @@ class B12xMoEWrapper:
         activation_precision: str = "fp4",
         quant_mode: Optional[str] = None,
         source_format: str = "modelopt",
-        swiglu_limit: Optional[float] = None,
         apply_router_weight_on_input: bool = False,
     ):
         r"""Configure the b12x fused-MoE wrapper.
@@ -346,7 +343,6 @@ class B12xMoEWrapper:
             self.quant_mode
         )
         self.source_format = source_format
-        self.swiglu_limit = swiglu_limit
         self.apply_router_weight_on_input = apply_router_weight_on_input
 
         # Pre-allocated objects. Both workspace slots may be populated so
@@ -568,7 +564,6 @@ class B12xMoEWrapper:
                     n=self.intermediate_size,
                     k=self.hidden_size,
                     activation_precision=self.activation_precision,
-                    activation=self.activation,
                 )
                 self._weight_key = weight_key
         else:
@@ -594,7 +589,6 @@ class B12xMoEWrapper:
             activation_precision=self.activation_precision,
             quant_mode=self.quant_mode,
             source_format=self.source_format,
-            swiglu_limit=self.swiglu_limit,
             apply_router_weight_on_input=self.apply_router_weight_on_input,
             _workspace=workspace,
             _weight_views=self._weight_views,
