+++ flashinfer/fused_moe/cute_dsl/blackwell_sm12x/moe_dynamic_kernel.py
@@ -60,33 +60,35 @@ from cutlass._mlir.dialects import llvm
 from cutlass.cute.nvgpu import cpasync
 from flashinfer.cute_dsl.fp4_common import (
     atomic_add_global_i32,
-    broadcast_f32_to_half2,
-    cp_async4_shared_global,
-    cp_async_u32_shared_global,
-    cp_async_u64_shared_global,
-    e2m1x8_mul_residual_to_e4m3x8,
-    e2m1x8_to_e4m3x8,
     fabs_f32,
     fmax_f32,
     fp8_e4m3_to_f32,
     ld_shared_f32,
     ld_shared_i32_relaxed,
-    ld_shared_u32,
-    atomic_add_shared_i32,
-    mxfp8_mma_m16n8k32_f32_e4m3,
-    quantize_block_fp4,
     quantize_block_fp4_fast,
-    quantize_block_fp8_mx,
     get_ptr_as_int64,
     st_global_f32,
     st_global_i32,
     shared_ptr_to_u32,
     st_shared_f32,
     st_shared_u8,
-    st_shared_u32,
     st_global_u64,
     st_global_v4_u32,
     warp_reduce,
+)
+from flashinfer.fused_moe.cute_dsl.utils import (
+    broadcast_f32_to_half2,
+    cp_async4_shared_global,
+    cp_async_u32_shared_global,
+    cp_async_u64_shared_global,
+    e2m1x8_mul_residual_to_e4m3x8,
+    e2m1x8_to_e4m3x8,
+    ld_shared_u32,
+    atomic_add_shared_i32,
+    mxfp8_mma_m16n8k32_f32_e4m3,
+    quantize_block_fp4,
+    quantize_block_fp8_mx,
+    st_shared_u32,
 )
 from flashinfer.gemm.kernels.dense_blockscaled_gemm_sm120_b12x import (
     Sm120B12xBlockScaledDenseGemmKernel as DenseGemmKernel,
