--- flashinfer/fused_moe/cute_dsl/blackwell_sm12x/w4a8/__init__.py
@@ -1,1 +1,0 @@
-"""W4A8 throughput-tier MoE pipeline (QMMA grouped GEMM in the W4A16 shape)."""
