+++ tests/moe/test_b12x_fused_moe.py
@@ -1043,190 +1043,6 @@ def create_relu2_moe_tensors(
 
 
 # =============================================================================
-# SwiGLU-OAI helpers (gated, gate-first "w31" source layout)
-# =============================================================================
-# Confirmed against b12x/b12x/moe/fused/reference.py::_apply_gated_activation
-# and b12x/tests/test_tp_moe_swigluoai_reference.py:
-#   gate = clamp(gate, max=limit)              # upper clamp only
-#   up   = clamp(up, min=-limit, max=limit)    # symmetric clamp
-#   out  = gate * sigmoid(alpha * gate) * (up + beta)
-# Defaults: limit=7.0, alpha=1.702, beta=1.0.
-SWIGLUOAI_DEFAULT_LIMIT = 7.0
-SWIGLUOAI_DEFAULT_ALPHA = 1.702
-SWIGLUOAI_DEFAULT_BETA = 1.0
-
-
-def swigluoai_activation(
-    gate: torch.Tensor,
-    up: torch.Tensor,
-    *,
-    limit: float = SWIGLUOAI_DEFAULT_LIMIT,
-    alpha: float = SWIGLUOAI_DEFAULT_ALPHA,
-    beta: float = SWIGLUOAI_DEFAULT_BETA,
-) -> torch.Tensor:
-    """SwiGLU-OAI gated activation (gate-first split)."""
-    g = torch.clamp(gate, max=float(limit))
-    u = torch.clamp(up, min=-float(limit), max=float(limit))
-    return g * torch.sigmoid(float(alpha) * g) * (u + float(beta))
-
-
-def compute_reference_moe_swigluoai(
-    hidden_states: torch.Tensor,
-    gemm1_weights_w31: torch.Tensor,
-    gemm2_weights: torch.Tensor,
-    token_selected_experts: torch.Tensor,
-    token_final_scales: torch.Tensor,
-    num_tokens: int,
-    num_experts: int,
-    top_k: int,
-    hidden_size: int,
-    intermediate_size: int,
-    fc2_input_scale: Optional[torch.Tensor],
-    *,
-    limit: float = SWIGLUOAI_DEFAULT_LIMIT,
-    alpha: float = SWIGLUOAI_DEFAULT_ALPHA,
-    beta: float = SWIGLUOAI_DEFAULT_BETA,
-) -> torch.Tensor:
-    """Reference MoE with SwiGLU-OAI activation.
-
-    ``gemm1_weights_w31`` is gate-first: rows [0:n] are gate, [n:2n] are up.
-    """
-    device = hidden_states.device
-    hidden_states = hidden_states.float()
-    gemm1_weights_w31 = gemm1_weights_w31.float()
-    gemm2_weights = gemm2_weights.float()
-    output = torch.zeros((num_tokens, hidden_size), dtype=torch.float32, device=device)
-
-    for token_idx in range(num_tokens):
-        token_input = hidden_states[token_idx : token_idx + 1]
-        for k in range(top_k):
-            expert_idx = token_selected_experts[token_idx, k].item()
-            scale = token_final_scales[token_idx, k].item()
-            if expert_idx < 0 or expert_idx >= num_experts:
-                continue
-            w1 = gemm1_weights_w31[expert_idx]
-            gemm1_out = token_input @ w1.T
-            gate = gemm1_out[:, :intermediate_size]
-            up = gemm1_out[:, intermediate_size:]
-            act = swigluoai_activation(gate, up, limit=limit, alpha=alpha, beta=beta)
-            if fc2_input_scale is not None:
-                act = quant_dequant_fp4_reference(act, fc2_input_scale, sf_vec_size=16)
-            w2 = gemm2_weights[expert_idx]
-            output[token_idx] += scale * (act @ w2.T).squeeze(0)
-    return output
-
-
-def create_swigluoai_moe_tensors(
-    num_tokens: int,
-    hidden_size: int,
-    intermediate_size: int,
-    num_experts: int,
-    num_local_experts: int,
-    top_k: int,
-    device: str = "cuda",
-    seed: int = 42,
-):
-    """Create gated MoE tensors with FC1 in gate-first ("w31") source order.
-
-    The dispatch flips w31 -> kernel-native [up; gate] in _get_weight_views;
-    the reference reads gate=rows[:n], up=rows[n:2n].
-    """
-    from flashinfer.fp4_quantization import fp4_quantize
-    from flashinfer.cute_dsl.utils import convert_sf_to_mma_layout
-
-    torch.manual_seed(seed)
-    sf_vec_size = 16
-
-    x_bf16 = (
-        torch.randn(num_tokens, hidden_size, dtype=torch.bfloat16, device=device) / 10
-    )
-    a1_gs = torch.tensor([1.0], device=device, dtype=torch.float32)
-    x_quantized, x_sf = fp4_quantize(
-        x_bf16, global_scale=a1_gs, sf_vec_size=sf_vec_size, is_sf_swizzled_layout=False
-    )
-    x_sf = x_sf.unsqueeze(-1)
-
-    router_logits = torch.randn(num_tokens, num_experts, device=device)
-    routing_weights = F.softmax(router_logits, dim=1, dtype=torch.float)
-    routing_weights, selected_experts = torch.topk(routing_weights, top_k, dim=-1)
-    routing_weights = routing_weights / routing_weights.sum(dim=-1, keepdim=True)
-    routing_weights = routing_weights.float()
-    selected_experts = selected_experts.to(torch.int32)
-
-    # FC1 weights in gate-first ("w31") order: rows [0:n] gate, [n:2n] up.
-    w1_bf16_w31 = (
-        torch.randn(
-            num_local_experts,
-            2 * intermediate_size,
-            hidden_size,
-            dtype=torch.bfloat16,
-            device=device,
-        )
-        / 10
-    )
-    w1_gs = torch.tensor([1.0], device=device, dtype=torch.float32)
-    w1_flat = w1_bf16_w31.reshape(
-        num_local_experts * 2 * intermediate_size, hidden_size
-    )
-    w1_q_flat, w1_sf_flat = fp4_quantize(
-        w1_flat, global_scale=w1_gs, sf_vec_size=sf_vec_size, is_sf_swizzled_layout=True
-    )
-    w1_q = w1_q_flat.view(num_local_experts, 2 * intermediate_size, hidden_size // 2)
-    w1_weight_sf = convert_sf_to_mma_layout(
-        w1_sf_flat,
-        m=2 * intermediate_size,
-        k=hidden_size,
-        num_groups=num_local_experts,
-        sf_vec_size=sf_vec_size,
-    )
-    w1_alpha = torch.ones(num_local_experts, device=device, dtype=torch.float32)
-
-    w2_bf16 = (
-        torch.randn(
-            num_local_experts,
-            hidden_size,
-            intermediate_size,
-            dtype=torch.bfloat16,
-            device=device,
-        )
-        / 10
-    )
-    w2_gs = torch.tensor([1.0], device=device, dtype=torch.float32)
-    w2_flat = w2_bf16.view(num_local_experts * hidden_size, intermediate_size)
-    w2_q_flat, w2_sf_flat = fp4_quantize(
-        w2_flat, global_scale=w2_gs, sf_vec_size=sf_vec_size, is_sf_swizzled_layout=True
-    )
-    w2_q = w2_q_flat.view(num_local_experts, hidden_size, intermediate_size // 2)
-    w2_weight_sf = convert_sf_to_mma_layout(
-        w2_sf_flat,
-        m=hidden_size,
-        k=intermediate_size,
-        num_groups=num_local_experts,
-        sf_vec_size=sf_vec_size,
-    )
-    w2_alpha = torch.ones(num_local_experts, device=device, dtype=torch.float32)
-    fc2_input_scale = torch.tensor([1.0], device=device, dtype=torch.float32)
-
-    return {
-        "x": x_quantized,
-        "x_sf": x_sf,
-        "x_bf16": x_bf16,
-        "token_selected_experts": selected_experts,
-        "token_final_scales": routing_weights,
-        "w1_weight": w1_q,
-        "w1_weight_sf": w1_weight_sf,
-        # gate-first bf16 reference weights (rows [:n]=gate, [n:2n]=up).
-        "w1_weight_bf16_w31": w1_bf16_w31,
-        "w1_alpha": w1_alpha,
-        "fc2_input_scale": fc2_input_scale,
-        "w2_weight": w2_q,
-        "w2_weight_sf": w2_weight_sf,
-        "w2_weight_bf16": w2_bf16,
-        "w2_alpha": w2_alpha,
-    }
-
-
-# =============================================================================
 # Test Class: Functional API (b12x_fused_moe)
 # =============================================================================
 
@@ -2885,509 +2701,6 @@ class TestRelu2Activation:
         torch.cuda.synchronize()
         assert not torch.isnan(output).any(), "NaN after ReLU2 CUDA graph replay"
         assert not (output == 0).all(), "All zeros after ReLU2 CUDA graph replay"
-
-
-# =============================================================================
-# Test Class: W4A8 throughput tier (w4a8_mx, fp4_e8m0_k32 source)
-# =============================================================================
-#
-# Pure-Torch MXFP4 weight generation + MXFP8xFP4 oracle, ported from the b12x
-# upstream test (tests/test_w4a8_mx_tp_moe.py + benchmarks/benchmark_ds4_moe.py
-# + b12x/moe/fused/reference.py). The b12x package is not importable from the
-# FlashInfer venv, so the self-contained pure-torch helpers are reproduced here
-# verbatim in numerics.
-
-_W4A8_FP4_MAG_LUT = (0.0, 0.5, 1.0, 1.5, 2.0, 3.0, 4.0, 6.0)
-_W4A8_FP4_LUT = (
-    0.0,
-    0.5,
-    1.0,
-    1.5,
-    2.0,
-    3.0,
-    4.0,
-    6.0,
-    -0.0,
-    -0.5,
-    -1.0,
-    -1.5,
-    -2.0,
-    -3.0,
-    -4.0,
-    -6.0,
-)
-_W4A8_FLOAT8_E4M3_MAX = 448.0
-_W4A8_INV_FLOAT8_E4M3_MAX = 1.0 / _W4A8_FLOAT8_E4M3_MAX
-
-
-def _w4a8_fp4_quantize_values(x: torch.Tensor) -> torch.Tensor:
-    """FP4 E2M1 quantization with kernel-matching tie-breaking."""
-    sign = torch.sign(x)
-    x = torch.abs(x.clone())
-    x[(x >= 0.0) & (x <= 0.25)] = 0.0
-    x[(x > 0.25) & (x < 0.75)] = 0.5
-    x[(x >= 0.75) & (x <= 1.25)] = 1.0
-    x[(x > 1.25) & (x < 1.75)] = 1.5
-    x[(x >= 1.75) & (x <= 2.5)] = 2.0
-    x[(x > 2.5) & (x < 3.5)] = 3.0
-    x[(x >= 3.5) & (x <= 5.0)] = 4.0
-    x[x > 5.0] = 6.0
-    return x * sign
-
-
-def _w4a8_fp4_encode_nibbles(values: torch.Tensor) -> torch.Tensor:
-    mags = values.abs()
-    idx = torch.zeros_like(values, dtype=torch.uint8)
-    for code, mag in enumerate(_W4A8_FP4_MAG_LUT):
-        idx = torch.where(mags == mag, torch.full_like(idx, code), idx)
-    sign_bit = (values < 0).to(torch.uint8) << 3
-    return idx | sign_bit
-
-
-def _w4a8_quantize_mxfp4_batched(
-    w: torch.Tensor,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """[E, rows, cols] float -> (packed fp4 [E, rows, cols//2] u8,
-    e8m0 scales [E, rows, cols//32] u8)."""
-    E, rows, cols = w.shape
-    blocked = w.view(E, rows, cols // 32, 32)
-    bmax = blocked.abs().amax(dim=-1, keepdim=True)
-    safe = torch.where(bmax > 0, bmax / 6.0, torch.ones_like(bmax))
-    exponent = torch.ceil(torch.log2(safe)).clamp(-127, 127)
-    byte = (
-        torch.where(bmax > 0, exponent + 127, torch.zeros_like(exponent))
-        .to(torch.uint8)
-        .squeeze(-1)
-        .contiguous()
-    )
-    scale = torch.where(bmax > 0, torch.exp2(exponent), torch.zeros_like(exponent))
-    q = _w4a8_fp4_quantize_values(
-        torch.where(
-            scale > 0, blocked / scale.clamp(min=1e-30), torch.zeros_like(blocked)
-        ).view(E, rows, cols)
-    )
-    nib = _w4a8_fp4_encode_nibbles(q)
-    pair = nib.view(E, rows, cols // 2, 2)
-    packed = (pair[..., 0] | (pair[..., 1] << 4)).contiguous()
-    return packed, byte
-
-
-def _w4a8_make_quantized_stack(
-    E: int, rows: int, cols: int, *, gen: torch.Generator, device: torch.device
-) -> tuple[torch.Tensor, torch.Tensor]:
-    w = torch.randn(E, rows, cols, generator=gen, device=device) * 0.05
-    return _w4a8_quantize_mxfp4_batched(w)
-
-
-def _w4a8_make_synthetic_mxfp4_moe(
-    E: int, k: int, n: int, *, seed: int, device: torch.device
-) -> dict:
-    """Kernel-order ([up; gate]) random mxfp4 expert weights + e8m0 grids."""
-    gen = torch.Generator(device=device)
-    gen.manual_seed(seed)
-    w13_fp4, w13_mx = _w4a8_make_quantized_stack(E, 2 * n, k, gen=gen, device=device)
-    w2_fp4, w2_mx = _w4a8_make_quantized_stack(E, k, n, gen=gen, device=device)
-    ones = torch.ones(E, dtype=torch.float32, device=device)
-    return {
-        "w13_fp4": w13_fp4,
-        "w13_mx": w13_mx,
-        "w2_fp4": w2_fp4,
-        "w2_mx": w2_mx,
-        "alphas": ones,
-        "input_scale": ones,
-    }
-
-
-def _w4a8_quant_dequant_mxfp8_rows(x: torch.Tensor) -> torch.Tensor:
-    """Per-32-block MXFP8 quantize-dequantize roundtrip (UE8M0 ceil block
-    scale, E4M3 RN-saturating payload) — matches the in-kernel numerics."""
-    orig_shape = x.shape
-    cols = orig_shape[-1]
-    blocked = x.to(torch.float32).reshape(-1, cols // 32, 32)
-    block_max = blocked.abs().amax(dim=-1, keepdim=True)
-    scale = block_max * _W4A8_INV_FLOAT8_E4M3_MAX
-    bits = scale.to(torch.float32).contiguous().view(torch.int32)
-    mant = bits & 0x007FFFFF
-    bumped = torch.where(mant != 0, (bits + 0x00800000) & 0x7F800000, bits)
-    rounded = bumped.view(torch.float32)
-    byte = ((bumped >> 23) & 0xFF).to(torch.uint8)
-    inv_bits = (254 - byte.to(torch.int32)).clamp(min=0) << 23
-    inv = inv_bits.view(torch.float32)
-    inv = torch.where(byte == 0, torch.zeros_like(inv), inv)
-    payload = (
-        (blocked * inv)
-        .clamp(-_W4A8_FLOAT8_E4M3_MAX, _W4A8_FLOAT8_E4M3_MAX)
-        .to(torch.float8_e4m3fn)
-        .to(torch.float32)
-    )
-    return (payload * rounded).reshape(orig_shape)
-
-
-def _w4a8_dequant_fp4(
-    packed_u8: torch.Tensor, rows: int, cols: int, fp4_lut: torch.Tensor
-) -> torch.Tensor:
-    if packed_u8.dtype == torch.int8:
-        packed_u8 = packed_u8.view(torch.uint8)
-    lo = (packed_u8 & 0x0F).to(torch.int64)
-    hi = ((packed_u8 >> 4) & 0x0F).to(torch.int64)
-    return torch.stack([fp4_lut[lo], fp4_lut[hi]], dim=-1).reshape(rows, cols)
-
-
-def _w4a8_effective_weight_e8m0_k32(
-    w_packed: torch.Tensor,
-    scale_bytes: torch.Tensor,
-    rows: int,
-    cols: int,
-    fp4_lut: torch.Tensor,
-) -> torch.Tensor:
-    raw = _w4a8_dequant_fp4(w_packed, rows, cols, fp4_lut)
-    scale = torch.exp2(scale_bytes.to(torch.float32) - 127.0)
-    n_blocks = cols // 32
-    return (raw.view(rows, n_blocks, 32) * scale[:rows, :n_blocks].unsqueeze(-1)).view(
-        rows, cols
-    )
-
-
-def _w4a8_moe_reference_mx(
-    x: torch.Tensor,
-    w1_fp4: torch.Tensor,
-    w1_mx_scales: torch.Tensor,
-    w1_alphas: torch.Tensor,
-    w2_fp4: torch.Tensor,
-    w2_mx_scales: torch.Tensor,
-    w2_alphas: torch.Tensor,
-    topk_ids: torch.Tensor,
-    topk_weights: torch.Tensor,
-    E: int,
-    K: int,
-    I_tp: int,
-) -> torch.Tensor:
-    """W4A8 MoE oracle (silu): MXFP8 activations x MXFP4 weights."""
-    device = x.device
-    fp4_lut = torch.tensor(_W4A8_FP4_LUT, dtype=torch.float32, device=device)
-    m = x.shape[0]
-    x_qd = _w4a8_quant_dequant_mxfp8_rows(x.float())
-    output = torch.zeros(m, K, dtype=torch.float32, device=device)
-    for eid in range(E):
-        route_mask = topk_ids == eid
-        token_mask = route_mask.any(dim=1)
-        if not bool(token_mask.any().item()):
-            continue
-        alpha_fc1 = float(w1_alphas[eid].item())
-        alpha_fc2 = float(w2_alphas[eid].item())
-        fc1_rows = w1_fp4.shape[1]
-        w13_eff = _w4a8_effective_weight_e8m0_k32(
-            w1_fp4[eid], w1_mx_scales[eid], fc1_rows, K, fp4_lut
-        )
-        w2_eff = _w4a8_effective_weight_e8m0_k32(
-            w2_fp4[eid], w2_mx_scales[eid], K, I_tp, fp4_lut
-        )
-        xs = x_qd[token_mask]
-        # Kernel weight order is [up; gate].
-        up_out = (xs @ w13_eff[:I_tp].T) * alpha_fc1
-        gate_out = (xs @ w13_eff[I_tp:].T) * alpha_fc1
-        intermediate = torch.sigmoid(gate_out) * gate_out * up_out
-        int_qd = _w4a8_quant_dequant_mxfp8_rows(intermediate)
-        down_out = (int_qd @ w2_eff.T) * alpha_fc2
-        route_weight = (topk_weights.float() * route_mask.float()).sum(dim=1)[
-            token_mask
-        ]
-        output[token_mask] += route_weight.unsqueeze(1) * down_out
-    return output
-
-
-@cute_dsl_available
-@sm120_required
-@cuda_13_required
-class TestW4A8Mx:
-    """W4A8 throughput-tier (quant_mode='w4a8_mx', fp4_e8m0_k32 source)."""
-
-    _E = 16
-    _K = 4096
-    _N = 1024
-    _TOPK = 4
-
-    def _routed_inputs(self, m: int, seed: int):
-        device = torch.device("cuda")
-        gen = torch.Generator(device=device)
-        gen.manual_seed(seed)
-        x = (torch.randn(m, self._K, generator=gen, device=device) * 2.0).to(
-            torch.bfloat16
-        )
-        logits = torch.randn(m, self._E, generator=gen, device=device)
-        topk_logits, topk_ids = torch.topk(logits, self._TOPK, dim=-1)
-        topk_weights = torch.softmax(topk_logits, dim=-1).float()
-        return x, topk_ids.to(torch.int32), topk_weights
-
-    def _run_tier(self, m: int, weights: dict, *, seed: int = 33):
-        from flashinfer.fused_moe.cute_dsl.blackwell_sm12x.moe_dispatch import (
-            launch_sm120_moe,
-        )
-
-        device = torch.device("cuda")
-        x, topk_ids, topk_weights = self._routed_inputs(m, seed)
-        out = torch.empty(m, self._K, dtype=torch.bfloat16, device=device)
-        launch_sm120_moe(
-            a=x,
-            topk_ids=topk_ids,
-            topk_weights=topk_weights,
-            w1_weight=weights["w13_fp4"],
-            w1_weight_sf=weights["w13_mx"],
-            w1_alpha=weights["alphas"],
-            w2_weight=weights["w2_fp4"],
-            w2_weight_sf=weights["w2_mx"],
-            w2_alpha=weights["alphas"],
-            num_experts=self._E,
-            top_k=self._TOPK,
-            num_local_experts=self._E,
-            scatter_output=out,
-            activation="silu",
-            quant_mode="w4a8_mx",
-            source_format="fp4_e8m0_k32",
-        )
-        torch.cuda.synchronize()
-        return out
-
-    def _tier_dispatched(self) -> bool:
-        from flashinfer.fused_moe.cute_dsl.blackwell_sm12x.moe_dispatch import (
-            _W4A8_TIER_WORKSPACE_CACHE,
-        )
-
-        return len(_W4A8_TIER_WORKSPACE_CACHE) > 0
-
-    def _oracle(self, m: int, weights: dict, *, seed: int = 33, n: int = None):
-        n = self._N if n is None else n
-        x, topk_ids, topk_weights = self._routed_inputs(m, seed)
-        return _w4a8_moe_reference_mx(
-            x.float(),
-            weights["w13_fp4"],
-            weights["w13_mx"],
-            weights["alphas"],
-            weights["w2_fp4"],
-            weights["w2_mx"],
-            weights["alphas"],
-            topk_ids,
-            topk_weights,
-            self._E,
-            self._K,
-            n,
-        )
-
-    def test_tier_band_matches_oracle(self):
-        """m=1024 x topk=4 = 4096 routed rows = the default tier floor: the
-        call must auto-dispatch to the throughput tier and match the oracle."""
-        from flashinfer.fused_moe.cute_dsl.blackwell_sm12x import moe_dispatch
-
-        moe_dispatch._W4A8_TIER_WORKSPACE_CACHE.clear()
-        moe_dispatch._W4A8_TIER_WEIGHT_CACHE.clear()
-        device = torch.device("cuda")
-        weights = _w4a8_make_synthetic_mxfp4_moe(
-            self._E, self._K, self._N, seed=21, device=device
-        )
-        m = 1024
-        out = self._run_tier(m, weights)
-        assert self._tier_dispatched(), (
-            "expected the w4a8 tier to serve 4096 routed rows"
-        )
-        ref = self._oracle(m, weights)
-        n_out = out.float().norm().item()
-        assert n_out > 0.01, f"tier output near-zero (norm={n_out})"
-        cos = torch.nn.functional.cosine_similarity(
-            out.float().flatten(), ref.float().flatten(), dim=0
-        ).item()
-        assert cos > 0.998, cos
-        n_ref = ref.float().norm().item()
-        assert 0.8 < n_out / n_ref < 1.25, (n_out, n_ref)
-
-    def test_prepared_hard_fail_relu2(self):
-        """A prepared-w4a8 object with an unsupported activation must raise a
-        RuntimeError matching 'cannot run' (the hard-fail contract)."""
-        import types
-
-        from flashinfer.fused_moe.cute_dsl.blackwell_sm12x.moe_dispatch import (
-            _maybe_launch_w4a8_tier,
-        )
-
-        device = torch.device("cuda")
-        m = 256
-        x, topk_ids, topk_weights = self._routed_inputs(m, 7)
-        out = torch.empty(m, self._K, dtype=torch.bfloat16, device=device)
-        # A sentinel "prepared" object — the refusal must trip before its
-        # fields are read.
-        prepared = types.SimpleNamespace(
-            w13_rp=None, w13_sfb=None, w2_rp=None, w2_sfb=None
-        )
-        with pytest.raises(RuntimeError, match="cannot run"):
-            _maybe_launch_w4a8_tier(
-                a=x,
-                w13_fp4=None,
-                w13_mx=None,
-                w1_alpha=None,
-                w2_fp4=None,
-                w2_mx=None,
-                w2_alpha=None,
-                topk_ids=topk_ids,
-                topk_weights=topk_weights,
-                output=out,
-                m=m,
-                k=self._K,
-                n=self._N,
-                weight_E=self._E,
-                num_topk=self._TOPK,
-                routed_rows=m * self._TOPK,
-                activation="relu2",
-                quant_mode="w4a8_mx",
-                apply_router_weight_on_input=False,
-                device=device,
-                prepared_w4a8=prepared,
-            )
-
-    def test_prepared_hard_fail_geometry(self):
-        """A prepared-w4a8 object with k % 256 != 0 must raise 'cannot run'."""
-        import types
-
-        from flashinfer.fused_moe.cute_dsl.blackwell_sm12x.moe_dispatch import (
-            _maybe_launch_w4a8_tier,
-        )
-
-        device = torch.device("cuda")
-        m = 256
-        _, topk_ids, topk_weights = self._routed_inputs(m, 9)
-        out = torch.empty(m, 4096, dtype=torch.bfloat16, device=device)
-        prepared = types.SimpleNamespace(
-            w13_rp=None, w13_sfb=None, w2_rp=None, w2_sfb=None
-        )
-        with pytest.raises(RuntimeError, match="cannot run"):
-            _maybe_launch_w4a8_tier(
-                a=torch.empty(m, 4000, dtype=torch.bfloat16, device=device),
-                w13_fp4=None,
-                w13_mx=None,
-                w1_alpha=None,
-                w2_fp4=None,
-                w2_mx=None,
-                w2_alpha=None,
-                topk_ids=topk_ids,
-                topk_weights=topk_weights,
-                output=out,
-                m=m,
-                k=4000,  # not a multiple of 256
-                n=self._N,
-                weight_E=self._E,
-                num_topk=self._TOPK,
-                routed_rows=m * self._TOPK,
-                activation="silu",
-                quant_mode="w4a8_mx",
-                apply_router_weight_on_input=False,
-                device=device,
-                prepared_w4a8=prepared,
-            )
-
-
-@cute_dsl_available
-@sm120_required
-@cuda_13_required
-class TestSwiGLUOAI:
-    """Tests for the SwiGLU-OAI gated activation (gate-first "w31" source).
-
-    SwiGLU-OAI is gated (FC1 = 2*intermediate) like silu, but the source FC1
-    weights arrive gate-first; the dispatch flips them to the kernel-native
-    [up; gate] order. Covers the micro (small m) and dynamic (large routed)
-    NVFP4 backends with the default and a non-default clamp limit.
-    """
-
-    def _run_and_check(
-        self,
-        *,
-        num_tokens: int,
-        top_k: int,
-        num_experts: int,
-        hidden_size: int,
-        intermediate_size: int,
-        swiglu_limit: Optional[float],
-    ):
-        from flashinfer import b12x_fused_moe
-
-        tensors = create_swigluoai_moe_tensors(
-            num_tokens=num_tokens,
-            hidden_size=hidden_size,
-            intermediate_size=intermediate_size,
-            num_experts=num_experts,
-            num_local_experts=num_experts,
-            top_k=top_k,
-        )
-
-        result = b12x_fused_moe(
-            x=tensors["x_bf16"],
-            w1_weight=tensors["w1_weight"],
-            w1_weight_sf=tensors["w1_weight_sf"],
-            w1_alpha=tensors["w1_alpha"],
-            fc2_input_scale=tensors["fc2_input_scale"],
-            w2_weight=tensors["w2_weight"],
-            w2_weight_sf=tensors["w2_weight_sf"],
-            w2_alpha=tensors["w2_alpha"],
-            token_selected_experts=tensors["token_selected_experts"],
-            token_final_scales=tensors["token_final_scales"],
-            num_experts=num_experts,
-            top_k=top_k,
-            activation="swigluoai_uninterleave",
-            quant_mode="nvfp4",
-            swiglu_limit=swiglu_limit,
-        )
-
-        assert result.shape == (num_tokens, hidden_size)
-        assert not torch.isnan(result).any()
-        assert not torch.isinf(result).any()
-
-        eff_limit = (
-            SWIGLUOAI_DEFAULT_LIMIT if swiglu_limit is None else float(swiglu_limit)
-        )
-        ref_output = compute_reference_moe_swigluoai(
-            hidden_states=tensors["x_bf16"].float().cuda(),
-            gemm1_weights_w31=tensors["w1_weight_bf16_w31"].float().cuda(),
-            gemm2_weights=tensors["w2_weight_bf16"].float().cuda(),
-            token_selected_experts=tensors["token_selected_experts"],
-            token_final_scales=tensors["token_final_scales"],
-            num_tokens=num_tokens,
-            num_experts=num_experts,
-            top_k=top_k,
-            hidden_size=hidden_size,
-            intermediate_size=intermediate_size,
-            fc2_input_scale=tensors["fc2_input_scale"],
-            limit=eff_limit,
-        )
-
-        passed, percent_within, atol = check_accuracy(result, ref_output)
-        assert passed, (
-            f"SwiGLU-OAI: {percent_within * 100:.2f}% within tolerance "
-            f"(atol={atol:.4f}, tokens={num_tokens}, top_k={top_k}, "
-            f"limit={eff_limit})"
-        )
-
-    @pytest.mark.parametrize("swiglu_limit", [None, 0.1])
-    @pytest.mark.parametrize("num_tokens", [1, 2, 4])
-    def test_swigluoai_micro_accuracy(self, num_tokens: int, swiglu_limit):
-        """Micro backend (small m) with default and non-default clamp limit."""
-        self._run_and_check(
-            num_tokens=num_tokens,
-            top_k=2,
-            num_experts=256,
-            hidden_size=256,
-            intermediate_size=512,
-            swiglu_limit=swiglu_limit,
-        )
-
-    @pytest.mark.parametrize("swiglu_limit", [None, 0.1])
-    @pytest.mark.parametrize("num_tokens", [128, 512])
-    def test_swigluoai_dynamic_accuracy(self, num_tokens: int, swiglu_limit):
-        """Dynamic backend (large routed) with default and non-default limit."""
-        self._run_and_check(
-            num_tokens=num_tokens,
-            top_k=2,
-            num_experts=256,
-            hidden_size=1024,
-            intermediate_size=2048,
-            swiglu_limit=swiglu_limit,
-        )
 
 
 if __name__ == "__main__":
